Refactor source directories and packages
