+++ bundles/model/org.eclipse.smarthome.model.thing.tests/src/test/java/org/eclipse/smarthome/model/thing/test/hue/TestHueChannelTypeProvider.java
